Experimental tests with xstatic reviews
This commit adds experimental jobs to run horizon with the master branches
of xstatic repositories. It allows us to test xstatic packages before
merging/releasing them.

The existing npm jobs do not support sibling installation into the tox
env "npm" which is used by the npm jobs. In a new experimental npm job
with xstatic master repos, sibling installation is now supported.

Note that this change is proposed only to the horizon repo
to avoid sending the similar change to many xstatic repos.

Change-Id: I5fa29c3af9c896561d1738471b7464a2ea81b7a0