Note on enmerkar in requirements.txt
I was asked if enmerkar is a run-time requirements or not in IRC.
It is not a run-time requirements of horizon. I considered moving
it to test-requirements.txt but horizon plugins depends on horizon
requirements to ensure enmerkar is installed when extracting
translation strings, so we keep it in requirements.txt.
It is worth having a note on this in requirements.txt.

Change-Id: I3e97bb413f2030607da93d892808719fcf0d4a5f