Refactor error messages.
Error messages can come from different projects(cinder, keystone,
nova, etc.),but we are currently catching all the exceptions and
return a generic error message and it hard for the user to determine
the source of error. Horizon team decided to add a collapse-able box
for error messages which shows the detailed error message on the
horizon UI during shanghai summit[1]. This patch do the same.

Partially-Implements blueprint refactor-error-messages

[1] https://etherpad.openstack.org/p/horizon-u-ptg#110

Change-Id: If0bd24540562b8f1330ac6cb7db5f1d354e1d1b7