Add custom SSL CA Cert support for api.cinder.get_microversion
This is a stable branch version of
commit 285c51f7e1b27a786fa9684abdc489998285b4e2 in the master branch.
custom SSL CA Cert support in python-cinderclient was introduced
in 5.0.0, but horizon train supports python-cinderclient >=4.0.1,
so we cannot backport it as-is.

This commit borrowed the logic of get_server_version() from
python-cinderclient 5.0.0 so that this can be backported to
older releases than train.

From the above reason, this is directly proposed to stable/train.

Change-Id: I776d2c8c6864067e7b44ce4b2f107c80b8b6d7fb
Closes-Bug: #1744670