Update pyScss to 1.3.7
pyScss 1.3.4 is not compatible with setuptools 46.
pyScss 1.3.7 was releaed to make it work with setuptools 46.

Note that a workaround for the material design theme
in commit ddc52a7405f0be38629f665d24bcf75022ce0018 is not backported
in this change as the workaround is a partial fix of bug 1771559.
Instead, a release note to mention this as a known issue is added.

python-memcached in test-requirements.txt is updated to
pass the requirements-check job.

Depends-On: https://review.opendev.org/716004
Change-Id: I336a3449c1ed8be3984da6366e9b45f8e4d3f5c0
(cherry picked from commit 2c44702724f25c8114e1a7318d45cd0f87a6c131)