Set tox_constraints_file for npm jobs and integration tests
npm jobs and integration tests always try to use the constraints
from releases.openstack.org and do not pick up the change in the gate
even when Depends-On is specified.

For npm jobs, https://review.opendev.org/#/c/709414/ allows us to pass
tox_constraints_file when creating the envvar of the constraints file.
This commit consumes it.

Change-Id: I396073182ff1e63df12ebb79c3ddd646277ddadf