Fix horizon-nodejs10-run jobs
Tox is not anymore pre-installed on the OpenDev CI images.
The file package.json needs tox, install it to fix the job.

Install it via a pre-run playbook so that this is run in addition to all
of the playbooks of the parent.

Change-Id: Icf82fa96f7d90453f78cf564c847682a286ed575
(cherry picked from commit 7cd0debebe1a300a96f8d0f21626793be0a0bc41)