Define periodic jobs
We sometimes create gate check changes but periodic jobs can do
most of them, particularly in stable branhces.
This commit is planed to backport to stable branches.
Note that we already have periodic-stable-jobs (for pyXX and docs jobs)
but it is nice to have npm and integration jobs in this coverage.
Once it is available in stable branches, I will configure the Grafana
dashboard to monitor healthiness in stable branches.

Note on stable backport:
In stable branches, we have a pipeline named 'periodic-stable'
and it is used in periodic-stable-jobs template.
It would be nice to use the same pipeline for all periodic jobs.

Change-Id: Ieb2fe4d2c4453e572c9251dc03c34d08bed62a0f
(cherry picked from commit 392bc2482eff851504b245c2e28de00694dd9eb8)