npm job: Add 'nodejs' bindep profile
During the focal migration, npm jobs in some plugins requires
extra packages like libfontconfig. They are specific to npm jobs,
so it would be nice to have a dedicated bindep profile (as 'test'
profile is too general for this purpose). horizon job definition
provides a common layer for the npm jobs, so we decide to add a
task to the job definition in the horizon repo. Each plugin can
add packages with 'nodejs' profile in its bindep.txt if needed.

Change-Id: I5eac6b1c724e9b21a78fda5ba41002224dd9b356