Add heat-dashboard cross-project tests
heat-dashboard is not active development now but it's still used
in a lot of deployments so it's good to check that at least
unit-tests works with the latest horizon.

python cross jobs now inherit openstack-tox-pyNN instead of
openstack-tox as openstack-tox-pyNN job has its own configuration
including node configuration and so on. It ensures the python
cross jobs are run in the same environment as they are run
in corresponding projects.

Co-Authored-By: Akihiro Motoki <amotoki@gmail.com>
Change-Id: Ibbb36743552295a13f6d9ff51f3ac2244d2e7f6d