Refactor error messages.
When we didn't pass the redirect parameter in exceptions.handle() then
'_alert_message.html' template is used to displays the error messages.
If details are empty in a case but it still shows the 'Details' href
which is wrong and this patch adds a condition in that template to fix
that issue.

This patch also fix review comment of change-id
5081359295e5311ca725e80efa0bc4b2ec811a0e and define a constant for
'\u2026' and use it in all files.

Note: '\u2026' is used for unicode glyph representing '...'.

Closes-Bug: #1893305

Partially-Implements blueprint refactor-error-messages

Change-Id: If4409a1245c983fb51712c6c817c11e884916694
Co-Author: Ivan Kolodyazhny <e0ne@e0ne.info>