Clarify the plan of OPENSTACK_NOVA_EXTENSIONS_BLACKLIST deprecation
This commit clarify the plan on the removal of
OPENSTACK_NOVA_EXTENSIONS_BLACKLIST setting.

We received feedbacks that SimpleTenantUsage nova API feature can
affect the nova API performance in larger deployments like [1].
Stopping the usage of SimpleTenantUsage in horizon is a good workaround
and it was the main motivation we introduced this setting.

Considering the situation that nova dropped the extension mechanism,
I believe the right way is to introduce a new setting which controls
the usage of SimpleTenantUsage rather than keeping the blacklist setting.

The current deprecation notice does not explain the migration path.
This commit clarifies the current plan.

[1] https://review.opendev.org/#/c/743600/

Change-Id: I10fdbf23b56745b1be6497532530dd394c62ff29