Fix ironic-ui cross job
ironic-ui npm job uses phantomjs which requires libfontconfig1.
libfontconfig1 is installed via bindep.txt in ironic-ui [1],
but in case of horizon it is not a direct dependency of horizon.
To handle the situation, this commit adds a playbook to install
libfontconfig1 and use it in horizon-cross-ironic-ui-npm job only
instead of adding libfontconfig1 to horizon bindep.txt.

[1] https://review.opendev.org/#/c/744708/

Change-Id: I525ab42d1d5a16b982f6afdfae0b6544a6c7a057