Add WEBROOT to ngdetails navigation
Default index URL for angular details page doesn't consider WEBROOT
while building navigations. That is why refreshing details page or
opening it in a new tab works perfectly when WEBROOT is "/" and fails
for apache deployment where WEBROOT is "/dashboard/".

This patch adds WEBROOT to index URL and builds a proper navigation.

Closes-Bug: #1902821

Change-Id: Id609e5499cc8f53dbae36febc945568e45d291f3