Update lower constraints and change browser to Firefox to unblock gate
Pip 20.x requires ensure_str method from six which was introduced
in v.1.12.0 and wrapt v.1.11 is required by astroid v.2.4.0 with
the pylint dependency. So this patch update six and wrapt version
in lower-constraints.txt to fix the gate.

It also updates the browser to Firefox in karma.conf.js and adds
'karma-firefox-launchers' in package.json because nodejs-job
fail with chromium-browser due to not able to access some URL.
It also removes 'karma-chrome-launcher' enteries from all files
as it is no longer used.  So let's switch to Firefox browser to
unblock the gate then we can investigate later how to fix it with
chromium-browser.

Depends-On: https://review.opendev.org/c/zuul/zuul-jobs/+/766744

Co-Authored-By: Vishal Manchanda <manchandavishal143@gmail.com>
Change-Id: I13e68d89f55407b1b3c64219969912700feef500
(cherry picked from commit 6f0200032d5fb26784c0ce1428ae0cad84ef058f)