Update devDependencies in package.json
This patch updates outdated JavaScript dev dependencies.

The latest Jasmine version makes nested beforeEach and unhandled
promise rejection fail, this was fixed.

Also, the latest Jasmine version runs tests in random order by default,
random is set to false for now to prevent tests failures.

Finally, the latest Jasmine version supports a new configuration option
failSpecWithNoExpectations. By default, this option is set to false and
Jasmine reports specs that have no expectations as passed. Setting this
option to true tells Jasmine to report specs without as expectations as
failed. For now, it is set to false for openstack-dashboard tests, this
should be changed in future.

Change-Id: I04d5c3ce7f7e0a9a38251e975b68bfa1998e556f