Cap bandit for python 2.7 env
bandit 1.6.3 supports only python>=3.5 and rokcy jobs with python 2.7
are broken when bandit 1.6.3 is being installed.
bandit <1.6.3 supports python 2.7.
We have no constraint for test only libraries so we need to cap it.

This change is only required in rocky or older releases,
so it is directly proposed to stable/rocky.

doc/requirements.txt also needs to be updated to pass the requirements-check
job as global-requirements.txt in the requirements repo was updated
after the rocky release.

Depends-On: https://review.opendev.org/c/openstack/requirements/+/766426
Change-Id: Iae6c7abecd1aec9db00c14f29fd98dd1a8682084