Cap bandit for python 2.7 env
bandit 1.6.3 supports only python>=3.5 and train jobs with python 2.7
are broken when bandit 1.6.3 is being installed.
bandit <1.6.3 supports python 2.7.
We have no constraint for test only libraries so we need to cap it.

doc/requirements.txt and lower-constraints.txt also needs to be
updated to pass the requirements-check and lower-constarints-job
respectively.

Change-Id: Ia8c69501a1362cb08344aa50d00ed184e9f3d49f