Bump min requirement of xstatic-jquery to 1.12.x
As discussed in the last horizon team meeting, we plan to upgrade
jquery to jquery 3 series. Jquery 3 upgrade assumes we already use
jquery 1.12.x ro 2.2.x. We already use jquery 1.12.4.1 in the master
for long but our requirements file says >=1.8.2. This commit bumps
the xstatic-jquery minimum requirement to 1.12.4.1 to ensure
we use 1.12.x before upgrading jQuery to 3.

xstatic-jquery-ui min version is also bumped to 1.12.0.1.
The current requirement is >=1.10.4 but we use 1.12.x series for long
and use 1.12.1.1 since 2018, so I believe it is safe to bump it.

Change-Id: I2a1ff5a4df5e278f89fcd73c26f1d8527b9c4d7f