Pin tempest and tempest-horizon in stein
stable/stein is in the extended maintenance phase and we will pin
tempest version to the latest tempest release (26.0.0) in Wallaby.
tempest 26.0.0 was released before merging tempest-horizon back,
so we need tempest-horizon too. This commit tries to install
tempest 26.0.0 and tempest-horizon 1.1.0 (latest release).
It makes the merge of tempest-horizon more smooth.

Change-Id: I5f2369e47cfaa9ccfdf8aa1f91348349c55b8515