Add openstack_auth to translation-related commands
openstack_auth was missing in pull_catalog and extract_messages
commands in manage.py. It is just because we forgot add it when
openstack_auth was merged from django-openstack-auth into horizon.

update_catalog did not run with python3 as write_po assumes fileobj is
opened as binary mode. We need to open a POT file with binary mode.

Change-Id: Ifa1c07c46d1081a299b030e523713e73b41faf19