Support Django 3.2 support (2)
In Django 3.2 CookieStorage stores messages in the RFC 6265 compliant
format [1][2]. This means that horizon messages pass through cookies are
encoded in a different way that messages are encrypted.

Previously horizon UT interpretes messages in cookies literally in its own way,
but the change on CookieStorage in Django 3.2 broke it. Horizon should not
depend on its own way. The suggested way I believe is to use a method defined
in django.contrib.messages.storage.cookie.CookieStorage.

[1] https://docs.djangoproject.com/en/3.2/releases/3.2/#miscellaneous
[2] https://github.com/django/django/commit/2d6179c819010f6a9d00835d5893c4593c0b85a0

Change-Id: I3e5e12265d0bc7b753bbe1f57acdd663b9dd3587