Support Django 3.2 support (3)
In Django 3.2, assertQuerysetEqual started to compare querystes directly.
As the migration check it checks the first element of the compared items
have a same type. "values()" method of OrderedDict does not return an object
which allows the index access in Python3, so this change breaks our UT.
We need to ensure to pass a list to assertQuerysetEqual.
The commit changes the related codes to ensure a list for such cases.

[1] https://github.com/django/django/pull/13557/commits/3f7b3275627385f8f7531fca01cdda50d4ec6b6e

Change-Id: If0a9d17c30de73a9870d0ebf3b3a82a14179c23d