Move "Domain" field under "User Name" and "Password"
Autofocus on the login screen now is on the first field. It's either
"User Name" or "Domain" with multidomain support. However it appears
to be more convenient to keep the focus on "User Name" whether there
is multidomain support or not. Also "Domain" is pre-filled with the
last domain used, so often there's no need to change it.

This patch moves "Domain" field under "User Name" and "Password" and
keeps autofocus always on "User Name".

Closes-Bug: #1916958

Change-Id: I590ce3a5cbaa0ebb470228a7114a60ff89467835
(cherry picked from commit b09e5b303c8ba0e967e5178794a39fcee628e4e0)