Fix community image handling in launch instance form
Community images which are not owned by a focused project are not
included in the normal image list API (i.e., without specifying
"community" as visibility), so we need an extra call of the image
list API specifying "community" visibility in a filter.

It was originally fixed by [1] but was dropped in [2].
The main motivation of [2] is to address "Duplicates in a repeater
are not allowed" error in the JavaScript console.

The approach of [2] was to drop multiple glance API calls, but
the right solution would be to check duplicated images included
in both API calls. This situation with duplicated images happens
when a community image is owned by a focused project. In such case,
the community image is included in the responses of both API calls.

This commit adds a logic to check image IDs already processed.

[1] https://review.opendev.org/c/openstack/horizon/+/614688
[2] https://review.opendev.org/c/openstack/horizon/+/640762

Closes-Bug: #1914045
Related-Bug: #1779250
Related-Bug: #1818508

Change-Id: I2ed1b6064ddd6f62818d6112e98e5d5a98beae9d
(cherry picked from commit c9bb0e95eaa7ae4f0cce21bc0a246a9b0cd2f880)