Fix `_data_table.html` weird conditional.
While troubleshooting a problem with pagination I noticed the
following conditional in `horizon/templates/horizon/common/_data_table.html`:

```
      {% if table.number_of_pages is defined %}
        {% include "horizon/common/_data_table_pagination.html" %}
      {% else %}
        {% include "horizon/common/_data_table_pagination_with_pages.html" %}
      {% endif %}
```

At first sight, it looks wrong. I mean, if there are
number_of_pages, we should use `_data_table_pagination_with_pages.html`.
However, this conditional solves to `True` if
table.number_of_pages is NOT defined, and then it solves to `False`,
if it is defined. It is a very weird (to me) behavior.

Therefore, I am suggesting to use a more readable and understandable
conditional in this patch.

We can simply use `if table.number_of_pages is None`, which will
behave as expected; if there are not number of pages, we use the
traditional paging, and then, if there are number of pages, we use
the paged data table approach.

Change-Id: I933b1f9399f40569c7661c1db5776d468917b966