Don't load user role assignment or groups tabs for non-admins
As a non admin user, navigate to Identity -> Users. Then click on the
username of your user to go to the detail page.

Only the allowed Overview tab is visible.

The view shows three tabs: Overview, Role assignments, Groups. Click on
either Role assignments or Groups. An error will appear, showing that
the API call is unauthorised, and the table content will fail to load.

This change fixes the issue by conditionally loading the tabs based on
policy.

Closes-Bug: #1920214
Change-Id: Ic8b723e6fd423b96a4f5eff54f9392cee534ed9e