Change with_data=False for swift_get_container
This function is only used in the Container get
method which retrieves details about a container.

It's not used for listing data in the container
even though with_data was True doing a GET instead
of a HEAD on the container.

This indirectly also fixes the Ceph bug [1] in the
Swift API implementation that doesn't expose metadata
correctly for a GET but does for a HEAD.

[1] https://tracker.ceph.com/issues/47861

Change-Id: Iba4215b2e3a414d1794faa05be31bf69e0e48bec
(cherry picked from commit c87f0e7a90d05f9355bd58165c25561dedae36d5)