integration tests: Relax router interface status check
Previously the router interface test in the integration tests
checked the status of an added router interface, but the initial
status varies depending on a network backend used in neutron.
In case of neutron L3 agent, a new router interface has "Down" status,
but in case of OVN a new router interface has "Active" status.
This caused the integration test failure after DevStack switched
the default network backend from ML2/OVS to ML2/OVN.

The initial status of a new router interface depends on a network
backend and the neutron API does not define the initial status,
so it is too strict for the horizon integration tests to check
the status. This commit drops the status check.
It is sufficient to check the existence of a new router interface.

Change-Id: Idfd4a9206e85558ba17e476a47b38ee9c1f566fa
Closes-Bug: #1930405