Make word wrapping consistent
Proposed patch enables word wrapping for 'Image name' and 'Host' columns
in affected dashboards. Currently, all potentially long names like instance
and image name are wrapped. Ellipsis might be a good idea, but it is not
implemented. Patch [1] just turns off wrapping for instance size (which is
reasonable, it is short in most cases) and hostname (might be very long in
prod environment, because in some cases backend generate some random name
for compute VM)

[1] https://review.opendev.org/c/openstack/horizon/+/5831

Related-Bug: #964738
Change-Id: I3058825c0e4ae114fb397670dbed1c1c6e983552